refactor: refactored root namespace
