+++ src/backend/SecretsProvider/SecretsProvider.csproj
@@ -9,8 +9,8 @@
         <GenerateAssemblyFileVersionAttribute>false</GenerateAssemblyFileVersionAttribute>
         <GenerateAssemblyCompanyAttribute>false</GenerateAssemblyCompanyAttribute>
         <GenerateAssemblyInformationalVersionAttribute>false</GenerateAssemblyInformationalVersionAttribute>
-        <RootNamespace>Ralfred</RootNamespace>
-        <AssemblyName>Ralfred</AssemblyName>
+        <RootNamespace>Ralfred.SecretsProvider</RootNamespace>
+        <AssemblyName>Ralfred.SecretsProvider</AssemblyName>
         <Nullable>enable</Nullable>
     </PropertyGroup>
