+++ src/backend/SecretsProvider/Startup.cs
@@ -1,13 +1,12 @@
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
 using Ralfred.Common.DependencyInjection;
 
 
-namespace Ralfred
+namespace Ralfred.SecretsProvider
 {
 	public class Startup
 	{
